Add AI difficulty settings and slider
The computer player now has three difficulty settings, 'Easy', 'Medium', and 'Hard'.
On 'Hard', the computer almost always makes the best move possible.
On 'Medium', move precedence is slightly tweaked to decrease chance of selecting a game-winning move. This is done by adding all moves returned by the decision functions to an array and choosing one at random.
On 'Easy', a significant number of random open spaces are pushed into the 'Medium' array and a move is randomly selected by array index.

These settings are controlled by a 3-point slider on the Main Menu. The input of this slider is used in a switch statement of the 'ai' script.